Allow multiple DigiAssets to same address
There is no reason that multiple outputs can't be sent to the same address other then there where few reasons to do it.   I propose for DigiAssets we allow many outputs to the same address so each can have a different asset.

Yes I do realise this only works if the assets are added first.  Will still throw an error if funds are added then try to add an asset.  This is acceptable limitation.